Fixes unit tests (#32)
Updated package.json to include the necessary dev dependencies for
testing (jasmine-node, q, nock). `jasmine` was removed as it is now
included in jasmine-node.

Running `jasmine-node` with tests in subdirectories wasn't correctly
running all of the tests. The fix for this was to move all of the tests
into the tests directory. All require paths were updated within the
tests to reflect the change.

Most of the failing tests were due to http vs https expectations and spy
calls expecting a callback function. These were fixed by using Jasmine's
`.any(Function)`.